Update to 10x performance with professional editing.
Based on the feedback from the prior pull request, I decided to hire a
professional editor to make sure this is a much higher quality. I think
this is much better than the previous version.

Also, my editor pointed out that the math works out to 15x performance
improvement, not 10x.